### 2021-03-13 #3100 @whedon
[PRE REVIEW]: ExaFMM: a high-performance fast multipole method library with C++ and Python interfaces
**Submitting author:** @tingyu66 (<a href="http://orcid.org/0000-0003-2520-0511">Tingyu Wang</a>)
**Repository:** <a href="https://github.com/exafmm/exafmm-t" target ="_blank">https://github.com/exafmm/exafmm-t</a>
**Version:** v0.1.1
