### 2021-03-15 #3100 @labarba
Possible reviewers:

- sivaramambikasaran is the first author of [HODLRlib](https://joss.theoj.org/papers/10.21105/joss.01167) in JOSS (H-matrices, a related method)
- pitsianis is the first author of [SG-t-SNE-Π](https://joss.theoj.org/papers/10.21105/joss.01577) in JOSS, and also a co-author of the [parallel AFMPB code](https://doi.org/10.1016/j.cpc.2014.12.022), targeting the same problem setting we have done with the code in our recent [arXiv preprint](https://arxiv.org/abs/2103.01048)
- the author or the other reviewers of [TBFMM](https://joss.theoj.org/papers/10.21105/joss.02444) in JOSS, another fast multipole method software
