### 2021-03-16 #3100 @poulson
Hi Berenger (@berenger-eu) :wave:.

Given your authorship of the TBFMM JOSS paper, please let me know if you would be willing to serve as a reviewer on this submission on a high-performance C++/Python FMM implementation.
