### 2021-03-16 #3100 @poulson
Hi Siva (@sivaramambikasaran) :wave:,

Given your expertise in FMM and its algebraic analogues, please let me know if you would be willing to serve as a reviewer for this submission on a high-performance C++/Python FMM implementation.
