### 2021-03-23 c2529e4
add copyright notice to intrinsics.h
