+++ _sources/api/variable_namespaceexafmm__t_1a57b203c2726ddd7089d526f0ef9be6c8.rst.txt
@@ -1,0 +1,13 @@
+.. _exhale_variable_namespaceexafmm__t_1a57b203c2726ddd7089d526f0ef9be6c8:
+
+Variable exafmm_t::dividerLength
+================================
+
+- Defined in :ref:`file_include_timer.h`
+
+
+Variable Documentation
+----------------------
+
+
+.. doxygenvariable:: exafmm_t::dividerLength
