+++ _sources/api/function_namespaceexafmm__t_1ae5d771aab7f5b200a29da6f1bdca073a.rst.txt
@@ -1,0 +1,13 @@
+.. _exhale_function_namespaceexafmm__t_1ae5d771aab7f5b200a29da6f1bdca073a:
+
+Template Function exafmm_t::build_tree(Body<T> \*, int, int, Body<T> \*, int, int, Node<T> \*, Nodes<T>&, NodePtrs<T>&, NodePtrs<T>&, FmmBase<T>&)
+==================================================================================================================================================
+
+- Defined in :ref:`file_include_build_non_adaptive_tree.h`
+
+
+Function Documentation
+----------------------
+
+
+.. doxygenfunction:: exafmm_t::build_tree(Body<T> *, int, int, Body<T> *, int, int, Node<T> *, Nodes<T>&, NodePtrs<T>&, NodePtrs<T>&, FmmBase<T>&)
