+++ _sources/api/file_include_exafmm_t.h.rst.txt
@@ -108,7 +108,7 @@ Enums
 -----
 
 
-- :ref:`exhale_enum_exafmm__t_8h_1a90321ace254beb0e9eacd5517ae3a4b4`
+- :ref:`exhale_enum_namespaceexafmm__t_1a90321ace254beb0e9eacd5517ae3a4b4`
 
 
 Defines
@@ -140,58 +140,58 @@ Typedefs
 --------
 
 
-- :ref:`exhale_typedef_exafmm__t_8h_1ad0c0d2f775528e2888aa7edada48e2de`
+- :ref:`exhale_typedef_namespaceexafmm__t_1ad0c0d2f775528e2888aa7edada48e2de`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1a4281ebcbc7f4a4e67086e45f2a15e313`
+- :ref:`exhale_typedef_namespaceexafmm__t_1a4281ebcbc7f4a4e67086e45f2a15e313`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1a44fbf7b196ac53a9de5e76ee42c6ccd5`
+- :ref:`exhale_typedef_namespaceexafmm__t_1a44fbf7b196ac53a9de5e76ee42c6ccd5`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1a320a500e2e4fde262073736f8bc3d64b`
+- :ref:`exhale_typedef_namespaceexafmm__t_1a320a500e2e4fde262073736f8bc3d64b`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1af47ea9bbea1f61c4d2ffec5f5f08ead8`
+- :ref:`exhale_typedef_namespaceexafmm__t_1af47ea9bbea1f61c4d2ffec5f5f08ead8`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1ae3ef4351bbc31781661c69d5a964ebf0`
+- :ref:`exhale_typedef_namespaceexafmm__t_1ae3ef4351bbc31781661c69d5a964ebf0`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1ab5d05a3cf636da5a84b86ba56f29a81f`
+- :ref:`exhale_typedef_namespaceexafmm__t_1ab5d05a3cf636da5a84b86ba56f29a81f`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1ae0423fc35c5984c4be74d6b8a0a47d29`
+- :ref:`exhale_typedef_namespaceexafmm__t_1ae0423fc35c5984c4be74d6b8a0a47d29`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1a3c3101bfaaaa5a02e9d0be11b8a0c69d`
+- :ref:`exhale_typedef_namespaceexafmm__t_1a3c3101bfaaaa5a02e9d0be11b8a0c69d`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1ab52b97a9c44134d8ef9aefeaff0bea0c`
+- :ref:`exhale_typedef_namespaceexafmm__t_1ab52b97a9c44134d8ef9aefeaff0bea0c`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1a2bea1ab2ea176e198342661fc44c5d09`
+- :ref:`exhale_typedef_namespaceexafmm__t_1a2bea1ab2ea176e198342661fc44c5d09`
 
-- :ref:`exhale_typedef_exafmm__t_8h_1a2d37b6036cc78f5b8cbb668b2f098b89`
+- :ref:`exhale_typedef_namespaceexafmm__t_1a2d37b6036cc78f5b8cbb668b2f098b89`
 
[…]
