+++ _sources/api/variable_namespaceexafmm__t_1a95cf5bec10c23f5ef1393ff69a3192cf.rst.txt
@@ -1,0 +1,13 @@
+.. _exhale_variable_namespaceexafmm__t_1a95cf5bec10c23f5ef1393ff69a3192cf:
+
+Variable exafmm_t::REL_COORD
+============================
+
+- Defined in :ref:`file_include_exafmm_t.h`
+
+
+Variable Documentation
+----------------------
+
+
+.. doxygenvariable:: exafmm_t::REL_COORD
