+++ _sources/api/function_namespaceexafmm__t_1a86c1b2909c9ba1c73e680684fd98bf82.rst.txt
@@ -1,0 +1,13 @@
+.. _exhale_function_namespaceexafmm__t_1a86c1b2909c9ba1c73e680684fd98bf82:
+
+Function exafmm_t::start
+========================
+
+- Defined in :ref:`file_include_timer.h`
+
+
+Function Documentation
+----------------------
+
+
+.. doxygenfunction:: exafmm_t::start(std::string)
