+++ _sources/api/function_build__tree_8h_1a960b9cf76e38f5647b229440cd189315.rst.txt
@@ -1,13 +1,0 @@
-.. _exhale_function_build__tree_8h_1a960b9cf76e38f5647b229440cd189315:
-
-Template Function exafmm_t::build_tree(Bodies<T>&, Bodies<T>&, NodePtrs<T>&, NodePtrs<T>&, FmmBase<T>&, const Keys&)
-====================================================================================================================
-
-- Defined in :ref:`file_include_build_tree.h`
-
-
-Function Documentation
-----------------------
-
-
-.. doxygenfunction:: exafmm_t::build_tree(Bodies<T>&, Bodies<T>&, NodePtrs<T>&, NodePtrs<T>&, FmmBase<T>&, const Keys&)
