+++ _sources/api/namespace_exafmm_t.rst.txt
@@ -80,13 +80,13 @@ Functions
 
 - :ref:`exhale_function_namespaceexafmm__t_1a4c5585ab36a873853bd1037a305174ff`
 
-- :ref:`exhale_function_namespaceexafmm__t_1aa9bbd6da308c547dd21851abe39449e4`
-
-- :ref:`exhale_function_namespaceexafmm__t_1a960b9cf76e38f5647b229440cd189315`
-
 - :ref:`exhale_function_namespaceexafmm__t_1ae5d771aab7f5b200a29da6f1bdca073a`
 
 - :ref:`exhale_function_namespaceexafmm__t_1a845cf95f32c6b703f8b35605eac17646`
+
+- :ref:`exhale_function_namespaceexafmm__t_1aa9bbd6da308c547dd21851abe39449e4`
+
+- :ref:`exhale_function_namespaceexafmm__t_1a960b9cf76e38f5647b229440cd189315`
 
 - :ref:`exhale_function_namespaceexafmm__t_1ac159c0760835de5ae0b738996fb1ee1b`
 
@@ -160,9 +160,9 @@ Functions
 
 - :ref:`exhale_function_namespaceexafmm__t_1a99a87ebc089fc38be60f0ba46a256485`
 
-- :ref:`exhale_function_namespaceexafmm__t_1a352c04e2f9ad1f9d382c02e8f4486d56`
-
 - :ref:`exhale_function_namespaceexafmm__t_1a71b1b9efc409aed0bb812fe16448a407`
+
+- :ref:`exhale_function_namespaceexafmm__t_1a352c04e2f9ad1f9d382c02e8f4486d56`
 
 - :ref:`exhale_function_namespaceexafmm__t_1ab3369b6745a2cddefcfdf32f02619587`
 
