### 2021-03-24 #3100 @berenger-eu
@poulson If I am correct, some checkboxes should be on the first post, right? Should I do something to make them appearing? Maybe it will be automatically done after all the reviewers will be selected(?) Thanks.
