### 2021-03-24 c556879
remove tree balancing functions in build_tree
