+++ include/build_tree.h
@@ -81,7 +81,7 @@ namespace exafmm_t {
   void build_tree(Body<T>* sources, Body<T>* sources_buffer, int source_begin, int source_end, 
                   Body<T>* targets, Body<T>* targets_buffer, int target_begin, int target_end,
                   Node<T>* node, Nodes<T>& nodes, NodePtrs<T>& leafs, NodePtrs<T>& nonleafs,
-                  const Keys& leafkeys, FmmBase<T>& fmm, bool direction=false) {
+                  FmmBase<T>& fmm, bool direction=false) {
     //! Create a tree node
     node->idx = int(node-&nodes[0]);  // current node's index in nodes
     node->nsrcs = source_end - source_begin;
@@ -92,14 +92,7 @@ namespace exafmm_t {
     node->key = getKey(iX, node->level);
 
     //! If node is a leaf
-    bool is_leaf_key = 1;
-    if (!leafkeys.empty()) {  // when leafkeys is given (when balancing tree) 
-      std::set<uint64_t>::iterator it = leafkeys[node->level].find(node->key);
-      if (it == leafkeys[node->level].end()) {  // if current key is not a leaf key
-        is_leaf_key = 0;
-      }
-    }
-    if (node->nsrcs<=fmm.ncrit && node->ntrgs<=fmm.ncrit && is_leaf_key) {
+    if (node->nsrcs<=fmm.ncrit && node->ntrgs<=fmm.ncrit) {
       node->is_leaf = true;
       node->trg_value.resize(node->ntrgs*4, (T)(0.));   // initialize target result vector
       if (node->nsrcs || node->ntrgs)
@@ -115,23 +108,21 @@ namespace exafmm_t {
           targets_buffer[i].ibody = targets[i].ibody;
         }
       }
-      // Copy sources and targets' coords and values to leaf (only during 2:1 tree balancing)
-      if (!leafkeys.empty()) {
-        Body<T>* first_source = (direction ? sources_buffer : sources) + source_begin;
-        Body<T>* first_target = (direction ? targets_buffer : targets) + target_begin;
-        for (Body<T>* B=first_source; B<first_source+node->nsrcs; ++B) {
-          for (int d=0; d<3; ++d) {
-            node->src_coord.push_back(B->X[d]);
-          }
-          node->isrcs.push_back(B->ibody);
-          node->src_value.push_back(B->q);
+      // Copy sources and targets' coords and values to leaf
+      Body<T>* first_source = (direction ? sources_buffer : sources) + source_begin;
+      Body<T>* first_target = (direction ? targets_buffer : targets) + target_begin;
+      for (Body<T>* B=first_source; B<first_source+node->nsrcs; ++B) {
+        for (int d=0; d<3; ++d) {
+          node->src_coord.push_back(B->X[d]);
         }
-        for (Body<T>* B=first_target; B<first_target+node->ntrgs; ++B) {
-          for (int d=0; d<3; ++d) {
-            node->trg_coord.push_back(B->X[d]);
-          }
-          node->itrgs.push_back(B->ibody);
+        node->isrcs.push_back(B->ibody);
+        node->src_value.push_back(B->q);
+      }
+      for (Body<T>* B=first_target; B<first_target+node->ntrgs; ++B) {
+        for (int d=0; d<3; ++d) {
+          node->trg_coord.push_back(B->X[d]);
         }
+        node->itrgs.push_back(B->ibody);
       }
       return;
     }
@@ -160,8 +151,7 @@ namespace exafmm_t {
       child[c].level = node->level + 1;
       build_tree(sources_buffer, sources, source_offsets[c], source_offsets[c] + source_size[c],
[…]
