### 2021-04-01 ddbe239
fix fmm.depth in build_tree
