### 2021-04-03 83e8d59
rewrite list construction functions, remove tree balancing
