+++ include/build_list.h
@@ -52,15 +52,16 @@ namespace exafmm_t {
   }
 
   /**
-   * @brief Given the integer index of an octant and its depth, return the Hilbert
-   * index of the leaf that contains the octant.
+   * @brief Given the 3D index of an octant and its depth, return the key of
+   * the leaf that contains the octant. If such leaf does not exist, return the
+   * key of the original octant.
    *
    * @param iX Integer index of the octant.
    * @param level The level of the octant.
    *
    * @return Hilbert index with level offset.
    */
-	uint64_t findgnt(const ivec3& iX, int level, const unordered_set<uint64_t>& leaf_keys) {
+	uint64_t find_key(const ivec3& iX, int level, const unordered_set<uint64_t>& leaf_keys) {
     uint64_t orig_key = getKey(iX, level, true);
     uint64_t curr_key = orig_key;
 		while (level>0) {
@@ -101,62 +102,65 @@ namespace exafmm_t {
            (max_diff == sum_radius);
   }
 
+  /**
+   * @brief Build lists for P2P, P2L and M2P operators for a given node.
+   *
+   * @param node Node.
+   * @param nodes Tree.
+   * @param leaf_keys The set of all leaf keys.
+   * @param key2id The mapping from a node's key to its index in the tree.
+   */
   template <typename T>
-  void build_list(Node<T>* node, const unordered_set<uint64_t>& leaf_keys,
-                  const unordered_map<uint64_t, size_t>& key2id, Nodes<T>& nodes) {
-    set<Node<T>*> Uset, Wset, Xset;
+  void build_other_list(Node<T>* node, Nodes<T>& nodes,
+                        const unordered_set<uint64_t>& leaf_keys,
+                        const unordered_map<uint64_t, size_t>& key2id) {
+    set<Node<T>*> P2P_set, M2P_set, P2L_set;
     Node<T>* curr = node;
     if (curr->key != 0) {
       Node<T>* parent = curr->parent;
-      ivec3 minIdx = 0;
-      ivec3 maxIdx = 1 << node->level;
+      ivec3 min_iX = 0;
+      ivec3 max_iX = 1 << node->level;
       ivec3 curr_iX = get3DIndex(curr->key);
       ivec3 parent_iX = get3DIndex(parent->key); 
       // search in every direction
       for (int i=-2; i<4; i++) {
         for (int j=-2; j<4; j++) {
           for (int k=-2; k<4; k++) {
-            // Index3 tryPath( 2*path2Node(parGNodeIdx) + Index3(i,j,k) );
-            ivec3 tryPath;
-            tryPath[0] = i;
-            tryPath[1] = j;
-            tryPath[2] = k;
-            tryPath += parent_iX * 2;
-            if (tryPath >= minIdx && tryPath < maxIdx && tryPath != curr_iX) {	
-              // int resGNodeIdx = findgnt(depth(curGNodeIdx), tryPath);
-              uint64_t res_key = findgnt(tryPath, curr->level, leaf_keys);
+            ivec3 direction;
+            direction[0] = i;
+            direction[1] = j;
+            direction[2] = k;
+            direction += parent_iX * 2;
+            if (direction >= min_iX && direction < max_iX && direction != curr_iX) {	
+              uint64_t res_key = find_key(direction, curr->level, leaf_keys);
               bool adj = is_adjacent(res_key, curr->key);
               Node<T>* res = &nodes[key2id.at(res_key)];
               if (res->level < curr->level) {
                 if (adj) {
                   if (curr->is_leaf) {
-                    Uset.insert(res);
-                  } else {
-                    ;
+                    P2P_set.insert(res);
                   }
                 } else {
-                  Xset.insert(res);
[…]
