### 2021-04-05 b8feb28
parallelize list construction, call P2P instead of P2L and M2P when necessary
