+++ include/test.h
@@ -55,48 +55,6 @@ namespace exafmm_t {
     }
 
     //! Dummy M2L operator.
-    /*
-    void M2L(NodePtrs<T>& nonleafs) {
-#pragma omp parallel for schedule(dynamic)
-      for (size_t i=0; i<nonleafs.size(); ++i) {
-        Node<T>* trg_parent = nonleafs[i];
-        NodePtrs<T>& M2L_list = trg_parent->M2L_list;
-        for (size_t j=0; j<M2L_list.size(); ++j) {
-          Node<T>* src_parent = M2L_list[j];
-          if (src_parent) {
-            // find src_parent's relative position to trg_parent
-            // NodePtrs<T>::iterator
-            auto it = std::find(trg_parent->colleagues.begin(),
-                                trg_parent->colleagues.end(), src_parent);
-            assert(it != trg_parent->colleagues.end());   // src_parent has to be trg_parent's colleague
-            int colleague_idx = it - trg_parent->colleagues.begin();
-            ivec3 parent_rel_coord;
-            parent_rel_coord[0] = colleague_idx % 3 - 1;
-            parent_rel_coord[1] = (colleague_idx/3) % 3 - 1;
-            parent_rel_coord[2] = (colleague_idx/9) % 3 - 1;
-            for(int src_octant=0; src_octant<8; ++src_octant) {
-              Node<T>* src_child = src_parent->children[src_octant];
-              ivec3 src_child_coord = octant2coord(src_octant);
-              for(int trg_octant=0; trg_octant<8; ++trg_octant) {
-                Node<T>* trg_child = trg_parent->children[trg_octant];
-                ivec3 trg_child_coord = octant2coord(trg_octant);
-                if(src_child && trg_child) {
-                  ivec3 rel_coord = parent_rel_coord*2 + (src_child_coord - trg_child_coord) / 2;  // calculate relative coords between children
-                  bool is_colleague = true;
-                  for(int d=0; d<3; ++d) {
-                    if(rel_coord[d]>1 || rel_coord[d]<-1)
-                      is_colleague = false;
-                  }
-                  if(!is_colleague)  // perform M2L if they are not neighbors
-                    trg_child->dn_equiv[0] += src_child->up_equiv[0];
-                }
-              }
-            }
-          }
-        }
-      }
-    }*/
-
     void M2L(NodePtrs<T>& nonleafs) {
 #pragma omp parallel for schedule(dynamic)
       for (size_t i=0; i<nonleafs.size(); ++i) {
