### 2021-04-07 e32dc62
fix examples and python wrapper
