### 2021-04-08 #3145 @berenger-eu
Dear @tingyu66,

I have a single tiny remark concerning the paper, which is clear and answers all the questions I had.
Instead of citing "Blanchard, P., Bramas" for ScalFMM, I would suggest citing the SISC paper:
https://epubs.siam.org/doi/abs/10.1137/130915662

I was able to compile the code (using the dependencies already installed on my computer) and run the tests.
I looked at the CI and saw that it tests compilation by loading the same modules as the ones given in the documentation.
I was able to install the python package and run an example.
I looked at the source code, and it is clear and compact, as described in the paper.
Furthermore, I got similar performances.
From my side, the documentation is also fine.

Therefore, I have only one remark, and it concerns the authors:
I saw that a fourth person (bloodysin <stvgm1993@gmail.com>) participate in the development.
I would suggest to make sure it is fine if this person is not in the list of the authors (who is maybe an alias to one of the authors).

I have also a question that I ask because I am interested to know more, but that does not need to be addressed in the paper.
I saw on the TODO.md that you plan to use GPUs.
Do you already have an idea of how you will do it?
Do you plan to use a runtime system, maybe legion?

Thanks.
