### 2021-04-08 f470076
update travis configuration for mpi tests
