+++ include/partition.h
@@ -1,0 +1,178 @@
+#ifndef partition_h
+#define partition_h
+#include "alltoall.h"
+#include "hilbert.h"
+
+namespace exafmm_t {
+  /**
+   * @brief MPI_Allreduce local bounds to get global bounds.
+   *
+   * @tparam T Target's value type (real or complex).
+   * @param sources Vector of sources.
+   * @param targets Vector of targets.
+   * @param x0 Coordinates of the center of the global bounding box.
+   * @param r0 Radius of the bounding box.
+   */
+  template <typename T>
+  void allreduceBounds(const Bodies<T>& sources, const Bodies<T>& targets, vec3& x0, real_t& r0) {
+    vec3 localXmin, localXmax, globalXmin, globalXmax;
+    localXmin = sources[0].X;
+    localXmax = sources[0].X;
+    for (size_t b=0; b<sources.size(); b++) {
+      localXmin = min(sources[b].X, localXmin);
+      localXmax = max(sources[b].X, localXmax);
+    }
+    for (size_t b=0; b<targets.size(); b++) {
+      localXmin = min(targets[b].X, localXmin);
+      localXmax = max(targets[b].X, localXmax);
+    }
+    MPI_Allreduce(&localXmin[0], &globalXmin[0], 3, MPI_REAL_T, MPI_MIN, MPI_COMM_WORLD);
+    MPI_Allreduce(&localXmax[0], &globalXmax[0], 3, MPI_REAL_T, MPI_MAX, MPI_COMM_WORLD);
+    x0 = (globalXmax + globalXmin) / 2;
+    r0 = fmax(max(x0-globalXmin), max(globalXmax-x0));
+    r0 *= 1.00001;
+  }
+
+  /**
+   * @brief Radix sort based on key and return permutation index.
+   *
+   * @param key Vector of keys.
+   * @param value Vector of permutation indices (after sorting).
+   * @param size Size of the vector.
+   */
+  void radixsort(std::vector<int> & key, std::vector<int> & value, int size) {
+    const int bitStride = 8;
+    const int stride = 1 << bitStride;
+    const int mask = stride - 1;
+    int maxKey = 0;
+    int bucket[stride];
+    std::vector<int> buffer(size);
+    std::vector<int> permutation(size);
+    for (int i=0; i<size; i++)
+      if (key[i] > maxKey)
+        maxKey = key[i];
+    while (maxKey > 0) {
+      for (int i=0; i<stride; i++)
+        bucket[i] = 0;
+      for (int i=0; i<size; i++)
+        bucket[key[i] & mask]++;
+      for (int i=1; i<stride; i++)
+        bucket[i] += bucket[i-1];
+      for (int i=size-1; i>=0; i--)
+        permutation[i] = --bucket[key[i] & mask];
+      for (int i=0; i<size; i++)
+        buffer[permutation[i]] = value[i];
+      for (int i=0; i<size; i++)
+        value[i] = buffer[i];
+      for (int i=0; i<size; i++)
+        buffer[permutation[i]] = key[i];
+      for (int i=0; i<size; i++)
+        key[i] = buffer[i] >> bitStride;
+      maxKey >>= bitStride;
+    }
+  }
+
+  /**
+   * @brief Partition sources and targets into different ranks.
+   *
+   * @tparam T Target's value type (real or complex).
+   * @param sources Vector of sources.
+   * @param targets Vector of targets.
+   * @param x0 Coordinates of the center of the global bounding box.
+   * @param r0 Radius of the bounding box.
+   * @param offset Hilbert key bin for each rank. For example, the key of
[…]

[…]
