### 2021-04-09 #3145 @labarba
hi @berenger-eu 👋  — first, thank you for being a reviewer!
Answering your question about the author list: the user `bloodysin` was a master's student with Rio a while back, and even though you see a bunch of commits, this was all hacking that eventually got removed, or in other words, he did not contribute meaningfully to the final product being submitted for JOSS.

Regarding your questions on GPUs, this is not in the roadmap now, and no GPU code is being submitted for review.
