+++ include/mpi_utils.h
@@ -1,8 +1,14 @@
 #ifndef mpi_utils_h
 #define mpi_utils_h
+#include <iostream>
 #include <mpi.h>
 
 namespace exafmm_t {
+#if FLOAT
+  MPI_Datatype MPI_REAL_T = MPI_FLOAT;        //!< Floating point MPI type
+#else
+  MPI_Datatype MPI_REAL_T = MPI_DOUBLE;        //!< Floating point MPI type
+#endif
   int MPIRANK;      //!< Rank of MPI communicator
   int MPISIZE;      //!< Size of MPI communicator
   int EXTERNAL;     //!< Flag to indicate external MPI_Init/Finalize
@@ -16,6 +22,19 @@ namespace exafmm_t {
 
   void stopMPI() {
     if (!EXTERNAL) MPI_Finalize();
+  }
+
+  template<typename T>
+  void printMPI(T data) {
+    int size = sizeof(data);
+    std::vector<T> recv(MPISIZE);
+    MPI_Gather(&data, size, MPI_BYTE, &recv[0], size, MPI_BYTE, 0, MPI_COMM_WORLD);
+    if (MPIRANK == 0) {
+      for (int irank=0; irank<MPISIZE; irank++ ) {
+        std::cout << recv[irank] << " ";
+      }
+      std::cout << std::endl;
+    }
   }
 }
 #endif
