### 2021-04-14 de7b3a1
add write_nodes function for debugging
