### 2021-04-21 #3145 @pitsianis
@tingyu66, I have not been able to compile on macos and linux.
The first one makes the most progress with the call
```
./configure CXX=g++-10
```
but it fails with
```
checking fftw3.h usability... no
checking fftw3.h presence... no
checking for fftw3.h... no
checking for fftwf_malloc in user specified location... no
checking for fftw_malloc in user specified location... no
configure: error: cannot find fftwf library.

```
even though I have `fftw` installed with `brew`.

The linux attempt fails with
```
checking for C++ compiler vendor... unknown
checking for OpenMP flag of C++ compiler... unknown
configure: error: don't know how to enable OpenMP for C++
 /t/exafmm-t-master> g++ --version
g++ (Ubuntu 7.5.0-3ubuntu1~18.04) 7.5.0
Copyright (C) 2017 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```
