### 2021-04-21 #3145 @tingyu66
Hello @pitsianis, our code currently does not support MacOS because many macros in our configure script have a hard time checking libraries on MacOS.

However, it should work on Linux though. I tested many versions of g++ ranging from 5.4.x to 10.2.x. It seems that the configure script did not find your g++ as the C++ compiler. Did you use `./configure CXX=g++` on Linux?
