### 2021-04-22 #3145 @pitsianis
On `Ubuntu 18.04.5 LTS (GNU/Linux 4.15.0-118-generic x86_64)` the `./configure CXX=g++` stops with
```
checking for fftw3.h... yes
checking for fftwf_malloc in user specified location... no
checking for fftw_malloc in user specified location... no
configure: error: cannot find fftwf library.
```
