### 2021-04-22 #3145 @tingyu66
> On `Ubuntu 18.04.5 LTS (GNU/Linux 4.15.0-118-generic x86_64)` the `./configure CXX=g++` stops with
>
> ```
> checking for fftw3.h... yes
> checking for fftwf_malloc in user specified location... no
> checking for fftw_malloc in user specified location... no
> configure: error: cannot find fftwf library.
> ```

Exafmm-t requires the FFTW3 shared library in both single and double precision, so we suggest to install `libfftw3-dev` package, as described in our [documentation](https://exafmm.github.io/exafmm-t/compile.html):
```bash
apt-get update
apt-get -y install libfftw3-dev
```

If you want to build FFTW3 library from source, you probably need to compile separately for each precision if I remember correctly. Then add the directory of the shared libraries to `LIBRARY_PATH` and `LD_LIBRARY_PATH` environment variables.
