### 2021-05-16 bee9c51
cite ScalFMM's SISC paper as suggested by reviewers
