+++ paper.bib
@@ -243,6 +243,17 @@
 	pages = {2444},
 }
 
+@article{agullo2014task,
+  title={Task-based FMM for multicore architectures},
+  author={Agullo, Emmanuel and Bramas, B{\'e}renger and Coulaud, Olivier and Darve, Eric and Messner, Matthias and Takahashi, Toru},
+  journal={SIAM Journal on Scientific Computing},
+  volume={36},
+  number={1},
+  pages={C66--C93},
+  year={2014},
+  publisher={SIAM}
+}
+
 @misc{wangETal2021,
       title={High-productivity, high-performance workflow for virus-scale electrostatic simulations with {Bempp-Exafmm}}, 
       author={Tingyu Wang and Christopher D. Cooper and Timo Betcke and Lorena A. Barba},
