### 2021-05-16 0f1b22a
fixed invalid DOIs, update Bempp-cl JOSS citation
