+++ paper.bib
@@ -63,7 +63,7 @@
 	Author = {L. Greengard and V. Rokhlin},
 	Date-Added = {2008-02-15 18:43:55 +0000},
 	Date-Modified = {2009-10-04 16:55:26 -0400},
-	Doi = {10.1016/0021-9991},
+	Doi = {10.1016/0021-9991(87)90140-9},
 	Journal = {J. Comput.\ Phys.},
 	Keywords = {FMM},
 	Number = {2},
@@ -212,18 +212,17 @@
    title = {pybind11 -- Seamless operability between C++11 and Python}
 }
 
-
 @article{Betcke2021,
-	author = {Timo Betcke and Matthew W. Scroggs},
-	doi = {10.21105/joss.02879},
-	journal = {J. Open Source Software},
-	keywords = {bempp, JOSS},
-	number = {0},
-	publisher = {The Open Journal},
-	title = {Bempp-cl: A fast {Python} based just-in-time compiling boundary element library},
-	volume = {0},
-	note = {Review underway at https://github.com/openjournals/joss-reviews/issues/2879},
-	year = {2021}
+  doi = {10.21105/joss.02879},
+  url = {https://doi.org/10.21105/joss.02879},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {59},
+  pages = {2879},
+  author = {Timo Betcke and Matthew W. Scroggs},
+  title = {Bempp-cl: A fast Python based just-in-time compiling boundary element library.},
+  journal = {Journal of Open Source Software}
 }
 
 @article{bramasTBFMMGenericParallel2020,
@@ -244,14 +243,15 @@
 }
 
 @article{agullo2014task,
-  title={Task-based FMM for multicore architectures},
-  author={Agullo, Emmanuel and Bramas, B{\'e}renger and Coulaud, Olivier and Darve, Eric and Messner, Matthias and Takahashi, Toru},
-  journal={SIAM Journal on Scientific Computing},
-  volume={36},
-  number={1},
-  pages={C66--C93},
-  year={2014},
-  publisher={SIAM}
+  title = {Task-based FMM for multicore architectures},
+  author = {Agullo, Emmanuel and Bramas, B{\'e}renger and Coulaud, Olivier and Darve, Eric and Messner, Matthias and Takahashi, Toru},
+  journal = {SIAM Journal on Scientific Computing},
+  doi = {10.1137/130915662},
+  volume = {36},
+  number = {1},
+  pages = {C66--C93},
+  year = {2014},
+  publisher = {SIAM}
 }
 
 @misc{wangETal2021,
