### 2021-05-16 #3145 @tingyu66
Hi @poulson, I just fixed the invalid DOIs and updated one reference as suggested by @berenger-eu in the `master` branch.
