### 2021-05-20 #3145 @poulson
Thanks -- I noticed that the paper could use a round of edits and that is what has been delaying me. I am going to rebuild and respond with a round of suggestions shortly.
