### 2021-05-20 #3145 @poulson
Here is a first batch of suggestions (all before the 'Statement of Need' section):

I recommend changing

1.  "It is re-written with a new design, after multiple re-implementations of the same algorithm, over a decade of work in the research group."
to
    "Our current implementation is the most recent of many across a decade of work in our research group."

2. "for what is an intricate and difficult algorithm to implement"
to
    "for an intricate algorithm"

3. "The Python binding in this new version allows calling it from a Jupyter notebook, reducing barriers for adoption. It is also easy to extend, and faster or competitive with state-of-the art alternatives."
to
    "Our new header-only C/C++ implementation is easier to extend, still competitive with state-of-the-art implementations, and includes a pybind11 Python interface."

4. I recommend deleting the sentence "Together with the likes of Krylov iterative methods and the fast Fourier transform, FMM is considered one of the top 10 algorithms of the 20-th century."

5. I recommend combining the surrounding two sentences of (4) into
    "The fast multipole method (FMM) was introduced more than 30 years ago as a means of reducing the complexity of N-body problems from O(N^2) to O(N) using hierarchical approximations of long-range interactions."
There is no need for the parenthetical explanation of N since that is suggested by the "N-body" problem reference.

6. I recommend weaking the statement "Two variants of hierarchical N-body algorithms exist" to "two major variants", and perhaps even mentioning that structures matrix algorithms (such as H-matrices, etc.) are algebraic analogues of FMM and tree codes.

7. I recommend rewording the later sentence to "For example, the integral formulation of problems modeled by elliptic partial differential equations can be reinterpreted as N-body interactions."

8. "kernel-independent variant of the algorithm" to "kernel-independent variant of FMM"

9. We can now delete the sentence "ExaFMM is a header-only library written in C/C++ and provides a Python API using pybind11" into earlier statement." due to its incorporation above.
