### 2021-05-20 #3145 @poulson
A second round of suggested edits:

1. Move "The first version of ExaFMM focused on low-accuracy optimizations and implemented a dual-tree traversal. It was GPU-enabled using CUDA, distributed via MPI, and multi-threaded via OpenMP." into separate paragraph alongside "Despite these efforts, it has remained a challenge in the FMM community to have a well-established open-source software package analogous to FFTW for the fast Fourier transform."

2. Paragraph break at "The 'alpha' version of ExaFMM...poor reusability *and* maint
ainability."
[Delete "as is the case with various other codes in this field. Already it was the fourth implementation of the algorithm within our research group, and two more came about over the years before the version we present in this paper." as the former is unfair without citations and the latter is redundant with earlier discourse.]

3. Further delete the following section due to it being more appropriate for a blog post or extended footnote:
"As a bit of history of this project...and to increased scientific impact."
