+++ paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 ExaFMM is an open-source library for fast multipole algorithms, providing high-performance evaluation of N-body problems in three dimensions, with C++ and Python interfaces.
 This new implementation is the most recent of many across a decade of work in our research group.
-Our goal for all these years has been to produce reusable, standard code for what is an intricate and difficult algorithm to implement. 
+Our goal for all these years has been to produce reusable, standard code for this intricate algorithm. 
 The Python binding in this new version allows calling it from a Jupyter notebook, reducing the barriers for adoption.
 It is also easy to extend, and faster or competitive with state-of-the art alternatives.
 
