+++ paper.md
@@ -41,7 +41,7 @@ Two major variants of hierarchical N-body algorithms exist: treecodes and FMM.
 Both were originally developed for fast evaluation of the gravitational potential field, but now have found many applications in different fields.
 For example, the integral formulation of problems modeled by elliptic partial differential equations can be reinterpreted as N-body interactions.
 In this way, N-body algorithms are applicable to acoustics, electromagenetics, fluid dynamics, and more.
-The present version of ExaFMM implements the so-called kernel-independent variant of the algorithm [@yingKernelindependentAdaptiveFast2004].
+The present version of ExaFMM implements the so-called kernel-independent variant of FMM [@yingKernelindependentAdaptiveFast2004].
 It supports computing both the potential and forces of Laplace, low-frequency Helmholtz and modified Helmholtz (a.k.a. Yukawa) kernels.
 Users can add other non-oscillatory kernels with modest programming effort.
 ExaFMM is a header-only library written in C/C++ and provides a Python API using `pybind11` [@pybind11].
