### 2021-05-30 ae534d4
editor rec 2.2
https://github.com/openjournals/joss-reviews/issues/3145
