### 2021-05-30 1a9146f
editor rec 2.3
https://github.com/openjournals/joss-reviews/issues/3145
