### 2021-05-30 78b7d56
pasted here paragraph deleted from paper.md
deleted in commit 1a9146f due to JOSS editor's request
