+++ _sources/api/file_include_build_non_adaptive_tree.h.rst.txt
@@ -51,9 +51,9 @@ Functions
 ---------
 
 
-- :ref:`exhale_function_namespaceexafmm__t_1ae5d771aab7f5b200a29da6f1bdca073a`
-
 - :ref:`exhale_function_namespaceexafmm__t_1a845cf95f32c6b703f8b35605eac17646`
+
+- :ref:`exhale_function_namespaceexafmm__t_1ae5d771aab7f5b200a29da6f1bdca073a`
 
 - :ref:`exhale_function_namespaceexafmm__t_1a68ab50d01815bca7c8b44e79b4282596`
 
