### 2021-05-30 #3145 @poulson
A long footnote might make sense as well! I just think it is a lengthy aside that distracts from the conciseness that is typical of JOSS articles.
