### 2021-05-30 #3145 @labarba
> A long footnote might make sense as well! I just think it is a lengthy aside that distracts from the conciseness that is typical of JOSS articles.

That's OK. We'll find another place and time to document the story. (It wouldn't work as a footnote, as it had citations.)
