### 2021-05-30 #3145 @labarba
Thank you @arfon! It's exciting to be on the other side of the process 😊

Huge thanks to the handling editor, @poulson, and to the reviewers: @berenger-eu, @pitsianis — your expert eyes on our work makes us more confident, and we're glad to see this in JOSS!
